requirements.txt
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
simple-term-menu==1.6.6
six==1.17.0
tzdata==2025.2
watchdog==6.0.0
